+++ Server_1.2.4/dimm/home/index/IMAP/MailInfo.java
@@ -21,7 +21,7 @@ import java.io.IOException;
 
 public interface MailInfo
 {
-    public int getUID();
+    public long getUID();
     public String getMID() throws IOException;
     public String getFlags();
     public long getRFC822size();
